update movement and physics loop
keys apply force
detect collision before render

+++ src/Player.h
@@ -2,30 +2,29 @@
 #include <SDL.h>
 #include <SDL_image.h>
 
-class Player {
+class Player
+{
 
 public:
-
-    int x_pos;
-    int y_pos;
-    int x_vel;
-    int y_vel;
-    int x_accel;
-    int y_accel;
+    float x_pos;
+    float y_pos;
+    float x_vel;
+    float y_vel;
+    float x_accel;
+    float y_accel;
 
     int width;
     int height;
 
     bool isJumping;
 
-    SDL_Texture* player_texture;
+    SDL_Texture *player_texture;
 
-    Player(int x, int y, int w, int h, SDL_Texture* t);
+    Player(float x, float y, int w, int h, SDL_Texture *t);
     void updatePosition();
 
     void applyForces();
 
     void detectCollisions();
     bool isColliding(SDL_Rect r);
-
 };
